Rename sandbox worker manager for clarity
