+++ apps/src/pythonlab/sandbox/constants.ts
@@ -1,7 +1,7 @@
 import {MultiFileSource, ProjectFile} from '@cdo/apps/lab2/types';
 
 // Message types exchanged between pyodideSandboxManager.ts (running on studio.code.org)
-// and pyodideWorkerManager.ts (running in a hidden iframe on the isolated
+// and pyodideSandboxWorkerManager.ts (running in a hidden iframe on the isolated
 // pyodide-sandbox.preview.codeprojects.org origin). Defined here so the string values
 // can't drift between the two webpack bundles that share this contract.
 // Messages produced directly by pyodideWebWorker.ts (sysout, syserr, run_complete, etc,
