+++ apps/src/sites/studio/pages/pyodide_sandbox/show.js
@@ -1,1 +1,1 @@
-import '@cdo/apps/pythonlab/sandbox/pyodideWorkerManager';
+import '@cdo/apps/pythonlab/sandbox/pyodideSandboxWorkerManager';
